Shuffled everything around to separate Platform and UserData IO
