+++ Platform Modules/Editor/UserDataIO_Editor.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 64f096f03896746bb950336b42a977ad
+timeCreated: 1587369393
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
