+++ Platform Modules/Standalone/PlatformIO.cs.meta
