+++ Platform Modules/Standalone/UserDataIO.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: fde8fa1ae70b5495e882da826081242b
+timeCreated: 1598857770
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
