+++ Runtime/DataStorage/PlatformIOBase.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 3811c984285044146bc35068bd8f1c71
+timeCreated: 1598854942
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
