+++ Runtime/DataStorage/SystemIOWrapper.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 1d9aa8861276447189750c6ec174c743
+timeCreated: 1598855586
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
