+++ Runtime/DataStorage/UserDataIOBase.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 0ed1fed9f400240749f5818bf2b30344
+timeCreated: 1598854942
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
